requirements.txt
google-genai>=0.8.0
chromadb>=0.5.0
pdfplumber>=0.11.0
gradio>=4.40.0
pydantic>=2.7.0
python-dotenv>=1.0.0
